Move AllowedMentions to common
